Created utility functions.
A new file with utility functions concerning representations in base n are
given. Three utilities are presented. One is taken from the original file i.e. 
base. An other is the inverse, it will produce the number given a 
representation. The last will produce the following representation if one is
added to the number.