Extended View
It is now possible to play a game and review it later.

It should be mentioned that there is a small bug in this version. You can not show the empty board in a review. The reason for this is the fact that 0 is false in python, so django will think that it is not present.